backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
bcrypt==3.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
cryptography==46.0.1
DateTime==5.5
debugpy==1.8.16
decorator==5.2.1
ecdsa==0.19.1
executing==2.2.1
fastapi==0.117.1
fpdf==1.7.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keyboard==0.13.5
matplotlib-inline==0.1.7
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
parso==0.8.5
passlib==1.7.4
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
email-validator==2.1.0
Pygments==2.19.2
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2

pyzmq==27.0.2
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.48.0
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
wcwidth==0.2.13
zope.interface==7.2
